+++ src/main/java/co/bitshifted/appforge/syncro/Syncro.java
@@ -6,23 +6,21 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package co.bitshifted.xapps.syncro;
+package co.bitshifted.appforge.syncro;
 
-import co.bitshifted.xapps.syncro.http.DownloadHandler;
-import co.bitshifted.xapps.syncro.launch.LaunchArgs;
-import co.bitshifted.xapps.syncro.sync.FileDiffChecker;
-import co.bitshifted.xapps.syncro.http.SyncroHttpClient;
-import co.bitshifted.xapps.syncro.launch.ReleaseProcessor;
-import co.bitshifted.xapps.syncro.model.ReleaseEntry;
-import co.bitshifted.xapps.syncro.model.UpdateCheckStatus;
-import co.bitshifted.xapps.syncro.model.UpdateInfo;
+import co.bitshifted.appforge.syncro.http.DownloadHandler;
+import co.bitshifted.appforge.syncro.http.SyncroHttpClient;
+import co.bitshifted.appforge.syncro.launch.LaunchArgs;
+import co.bitshifted.appforge.syncro.launch.ReleaseProcessor;
+import co.bitshifted.appforge.syncro.model.ReleaseEntry;
+import co.bitshifted.appforge.syncro.model.UpdateCheckStatus;
+import co.bitshifted.appforge.syncro.model.UpdateInfo;
+import co.bitshifted.appforge.syncro.sync.FileDiffChecker;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.List;
