+++ src/main/java/co/bitshifted/appforge/syncro/SyncroUtils.java
@@ -6,14 +6,11 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package co.bitshifted.xapps.syncro;
+package co.bitshifted.appforge.syncro;
 
 import java.io.File;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.Locale;
-import java.util.Set;
 
 /**
  * @author Vladimir Djurovic
