+++ src/main/java/co/bitshifted/appforge/syncro/model/FileEntry.java
@@ -8,7 +8,7 @@
  *
  */
 
-package co.bitshifted.xapps.syncro.model;
+package co.bitshifted.appforge.syncro.model;
 
 import java.io.File;
 import java.nio.file.Path;
