+++ src/main/java/co/bitshifted/appforge/syncro/model/UpdateCheckStatus.java
@@ -6,7 +6,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package co.bitshifted.xapps.syncro.model;
+package co.bitshifted.appforge.syncro.model;
 
 public enum UpdateCheckStatus {
 	NO_UPDATE, UPDATE_AVAILABLE, ERROR;
